basic_example/requirements.txt
Flask 
SQLAlchemy 
openfga_sdk 
authlib>=1.0
requests>=2.27.1
python-dotenv>=0.19.2